Moved & simplified classes for serialisation
