+++ runlab/src/main/java/RunLab/Objects/Strava/Route.java
@@ -1,0 +1,5 @@
+package RunLab.Objects.Strava;
+
+public class Route {
+
+}
