+++ runlab/src/main/java/RunLab/Objects/Strava/Split.java
@@ -1,0 +1,14 @@
+package RunLab.Objects.Strava;
+
+public class Split {
+    private float distance;
+    private float elapsed_time;
+    private float elevation_difference;
+    private float moving_time;
+    private float split;
+    private float average_speed;
+    private float average_grade_adjusted_speed;
+    private float average_heartrate;
+    private float pace_zone;
+
+}
